+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `laravel-imgproxy` will be documented in this file.
 
+## v1.2.0
+
+- !!! make Imagor object immutable
+- test cases for image sizes
+
 ## v1.0.0 (IMAGOR / Sandstorm) - 2025-09-27
 
 Initial release of the Imagor version (forked off)